WIP before switching back from LXD to LXC.
The LXD version will likely work, but the LXD module
has a problem in their module in that they use salt.utils.fopen
instead of salt.utils.files.fopen.  I can fix it, but I also
have a problem with having to install openssh on all these instances
in order to use saltify.  I'm thinking it might be easier to just
go back to LXC.